app/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.0.1
requests==2.29.0
paho-mqtt==1.6.1
google-cloud-bigquery==3.0.1
Flask==2.2.3
Flask-MQTT==1.1.1
Flask-SocketIO==5.3.2
Flask-DebugToolbar